Fix unwanted Hugo module updates
We need to avoid root's go.mod and go.sum being modified. We therefore rename
wite.toml as tests/config/demo-site.toml, which Hugo does
not automatically use, and update our scripts to point to this config
for testing builds and the demo site. In addition, we make
`tests/config` the source (build) directory (which takes the build out
project root).

Signed-off-by: Daniel F. Dickinson <dfdpublic@wildtechgarden.ca>